refactor(console): split manual capture pipeline
